Moved tests around. There are now three files. One for simple methods, one for collections and one for object
